% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_factorALL.R
\encoding{UTF-8}
\name{as_factorALL}
\alias{as_factorALL}
\title{Convertir a factor todas las variables tipo caracter de un dataframe}
\usage{
as_factorALL(dataframe)
}
\arguments{
\item{dataframe}{Dataframe}
}
\value{
Un dataframe con todas las variables tipo caracter convertidas a factor.
}
\description{
Convierte a factor todas las columnas tipo caracter de un dataframe.
}
\examples{
data(iris)
iris2<- as_factorALL(iris)
str(iris2)
}
