% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_mfrow.R
\name{auto_mfrow}
\alias{auto_mfrow}
\title{Funcion para auto mfrow tomada de la paqueteria mjcgraphics}
\usage{
auto_mfrow(nplots, setup = TRUE)
}
\arguments{
\item{nplots}{an integer in 1 to 49}

\item{setup}{if \code{TRUE}, then the graphical parameters (par) is set-up if
\code{FALSE}, and nplots < the number of spaces for plots in the device, then}

\item{par}{par
blank plots are added to fill in the unused spaces.}
}
\description{
automaticamente elige cuantas columnas y filas se usaran para
mostrar las graficas \code{par(mfrow)}. creada por Mark Cowley, 3 June 2006
}
\keyword{internal}
