% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charSummary.R
\encoding{UTF-8}
\name{charSummary}
\alias{charSummary}
\title{Resumen de las variables caracter o factor de un dataframe}
\usage{
charSummary(df)
}
\arguments{
\item{df}{Una base de datos, dataframe.}
}
\value{
Un resumen del dataframe.
}
\description{
Esta es una función que permite obtener el resumen de todas las variables no numéricas de un
dataframe. Permite obtener el número de observaciones (n), número de observaciones faltantes (miss),
porcentaje de observaciones faltantes (miss\%), número de factores únicos (unique) y el top 5 de niveles.
Función tomada de la paquetería: xda: R package for exploratory data analysis.
}
\examples{
charSummary(iris)
}
