% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq.post.hoc.R
\encoding{UTF-8}
\name{chisq.post.hoc}
\alias{chisq.post.hoc}
\title{Contrastes post hoc de una matriz de datos con chi-cuadrada}
\usage{
chisq.post.hoc(
  tabla,
  test = c("fisher.test"),
  popsInRows = TRUE,
  control = c("fdr", "BH", "BY", "bonferroni", "holm", "hochberg", "hommel"),
  digits = 4,
  simulate.p.value = NULL
)
}
\arguments{
\item{tabla}{Una tabla o matriz usada para la función chisq.test}

\item{test}{Método usado para ajustar el valor de P.}

\item{popsInRows}{Un argumento lógico que indica si las poblaciones forman las filas de la tabla (por defecto =TRUE) o no (=FALSE).}

\item{control}{Indica el método de control a utilizar. fdr, BH, BY, bonferroni, holm, hochberg y hommel.}

\item{digits}{Número de digitos a mostrar.}

\item{simulate.p.value}{Si argumento es TRUE, indica que se calculan los valores p mediante la simulación de Monte Carlo y 5000 replicas. Para más información consulta la funcion 'chisq.test'.}
}
\value{
Contrastes 'post-hoc' de chi cuadrada para dos niveles usando Fisher.test.
}
\description{
Muestra las comparaciones post-hoc por pares de una prueba Chi cuadrada significativa. Calcula los contrastes si hay más de dos grupos. Se aplica una prueba de fisher individual para saber que grupos difieren. Los valores p ajustados pueden calcularse con una amplia variedad de métodos: fdr, BH, BY, bonferroni, holm, hochberg y hommel. Funciona básicamente como una función envolvente que envía los valores p "brutos" no ajustados de cada prueba de chi-cuadrado por pares a la función p.adjust en el programa R base. La función p.adjust debe ser consultada para una mayor descripción de los métodos utilizados. Función tomada y modificada de la paqueteria fifer. https://www.rdocumentation.org/packages/fifer/versions/1.1
}
\examples{
table <- as.table(rbind(c(14, 43), c(23, 47)))
rownames(table) <- c("Hembras", "Machos")
chisq.post.hoc(table,test = c("fisher.test"))

xtab <- as.table(rbind(c(180, 145), c(179, 106),c(510, 196), c(862, 23)))
dimnames(xtab) <- list(
Tratamiento = c("1st", "2nd", "3rd", "Control"),
Genero = c("hembra", "macho"))
chisq.post.hoc(xtab)
chisq.post.hoc(xtab,control = c("fdr"))
chisq.post.hoc(xtab,control = c("BY"))
chisq.post.hoc(xtab,control = c("bonferroni"))
chisq.post.hoc(xtab,control = c("bonferroni"), simulate.p.value = TRUE)
}
