% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion.plot.R
\encoding{UTF-8}
\name{dispersion.plot}
\alias{dispersion.plot}
\title{Graficos de dispersion de un modelo}
\usage{
dispersion.plot(modelo, type = "pearson", overdispersion.term = NULL)
}
\arguments{
\item{modelo}{Modelo con distribuci칩n Poisson, Quasipoisson, binomial negativo.}

\item{type}{Tipo "pearson" por default.}

\item{overdispersion.term}{Par치metro de sobredispersion.}
}
\value{
Gr치ficos de dispersion de un modelo.
}
\description{
Graficar la sobredispersion de un modelo poisson, binomial negativo y Quasipoison de la paqueteria lme4,
glmer, glmmTMB, etc. GLMM y GLM. M치s detalles en:  https://github.com/glmmTMB/glmmTMB/issues/224
}
\examples{
data<- warpbreaks
modelo <- glm(breaks ~ wool+tension, poisson, data= data)
dispersion.plot(modelo)
}
