% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escalar_datos.R
\encoding{UTF-8}
\name{escalar_datos}
\alias{escalar_datos}
\title{Estandarizar las variables de un dataframe}
\usage{
escalar_datos(df, variables = NULL)
}
\arguments{
\item{df}{Un dataframe}

\item{variables}{Variables dataframe que se van a escalar}
}
\value{
Dataframe escalado y centrado
}
\description{
Estandariza las variables de un dataframe mediante el centrado
y el escalado. Para cada valor de una variable, simplemente restamos el
valor medio de la variable, luego dividimos por la desviacion estandar de
la variable. Esto significa que vamos a escalar los valores de manera que
la variable tenga una media de 0 y una desviaion estandar de 1.
Esto es util cuando las variables explicativas tienen diferentes escalas,
y a menudo no contribuyen por igual al modelo. Por ejemplo, si los valores
de una variable oscilan entre 0 y 10000 y los valores de otra variable oscilan
entre 0 y 10, la variable con el intervalo mas grande recibira un mayor peso
en el analisis. Al estandarizar las variables, podemos estar seguros de que
cada variable contribuye por igual al an?lisis. Este procedimiento no
representa una transformacion de los datos, y no cambia la magnitud del efecto ni
la significancia de la variable en el modelo.
Ver la funcion "stdize" de MuMIn package para mas detalles.
}
\examples{
data(iris)
#Si no se especifica que variables excluir, se escalan todas las variables numericas del dataframe.
escalar_datos(iris)
escalar_datos(iris, )
#También se puede especificar que variables especificamente se van a escalar
escalar_datos(iris, variables = c("Petal.Width","Petal.Length"))
escalar_datos(iris, variables = c(1,3))
}
