% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fam_link_mer.R
\name{fam_link_mer}
\alias{fam_link_mer}
\title{Función para determinar la familia de distribución y la función de enlace en el objeto 'mer'}
\usage{
fam_link_mer(mod)
}
\arguments{
\item{mod}{Modelo}
}
\value{
familia de distribución y link function
}
\description{
función de utilidad para determinar la familia de distribución y la función de enlace en el objeto de la clase 'mer'.
}
\note{
Función tomada de la paquetería AICcmodavg.
}
\keyword{internal}
