% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpairs_dfnum.R
\encoding{UTF-8}
\name{ggpairs_dfnum}
\alias{ggpairs_dfnum}
\title{ggpairs_dfnum}
\usage{
ggpairs_dfnum(dataframe, var.response = NULL)
}
\arguments{
\item{dataframe}{Dataframe}

\item{var.response}{Si se proporciona el nombre de la variable de respuesta, esta se coloca al final, para mostrarla en el eje Y. Para facilitar la observación de la relación entre las variables explicativas y la variable de respuesta.}
}
\value{
Un gráfico de correlación con valores de correlación, valores de p y lineas de
regresión lineal y no lineal de cada relación.
}
\description{
Matriz de correlación basada en ggpairs con lineas de regresion lineal y no lineal. Matriz de correlación de variables numéricas. Esta función selecciona solo las variables numéricas y aplica la función ggpairs
de la paqueteria GGally.
}
\examples{
ggpairs_dfnum(iris)
ggpairs_dfnum(iris, var.response = "Petal.Length")
}
