% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_curva.R
\encoding{UTF-8}
\name{hist_curva}
\alias{hist_curva}
\title{Grafica un histograma de frecuencias y qqplot de una variable}
\usage{
hist_curva(variable)
}
\arguments{
\item{variable}{Columna o vector numérico.}
}
\value{
Un gráfico de correlación con valores de correlación, valores de p y lineas de
regresión lineal y no lineal de cada relación.
}
\description{
Crea un histograma de frecuencias con una linea de densidad de la distribución y una prueba de normalidad
. Además de mostrar un qqplot de la variable. Si la longitud del vector es < 30 aplica una prueba Shapiro
-Wilk, si es mayor usa ks.test.
}
\examples{
hist_curva(iris$Sepal.Width)
}
