% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiar_nombres.R
\docType{data}
\name{mu_to_u}
\alias{mu_to_u}
\title{Constante para ayudar a cambiar de mu a u}
\format{
An object of class \code{character} of length 10.
}
\usage{
mu_to_u
}
\description{
Se trata de un vector de caracteres con los nombres de todos los puntos de codigo Unicode
conocidos que se parecen al simbolo griego mu o al micro y a los valores de "u".  Esto es
destinado a simplificar el mapeo de mu o micro en Unicode al caracter "u" con \code{limpiar_nombres()} y \code{limpiar_nombres2()}.
}
\details{
Mirar la ayuda de \code{limpiar_nombres()} para saber como usarla.
}
\seealso{
Other Set names: 
\code{\link{limpiar_nombres}()}
}
\concept{Set names}
\keyword{internal}
