% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_ggplot_themes.R
\encoding{UTF-8}
\name{my_theme_ggplot}
\alias{my_theme_ggplot}
\title{Tema 1 de ggplot predefinidos por mi para publicaciones}
\usage{
my_theme_ggplot(axis.size = NULL, axis.title.size = NULL)
}
\arguments{
\item{axis.size}{Tamaño de los ejes.}

\item{axis.title.size}{Tamaño de la leyenda del eje.}
}
\value{
Un grafico de ggplot con el tema predefinido.
}
\description{
Usa parámetros de ggplot predefinidos para una publicación científica, se basa en mis criterios y gustos
personales. Se puede especificar tamaño de los ejes X y Y, negritas y sin fondo.
}
\examples{
#Para usar el tema predefinido
#require(ggplot2)
#data(iris)
#ggplot(iris, aes(x=Petal.Length, y=Petal.Width)) +
#geom_point() +
#geom_smooth(method=lm , color="red", fill="#69b3a2", se=TRUE) +
#my_theme_ggplot(axis.size= 10, axis.title.size= 12)
}
