% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_ggplot_themes2.R
\encoding{UTF-8}
\name{my_theme_ggplot2}
\alias{my_theme_ggplot2}
\title{Tema2  de ggplot predefinidos por mi para publicaciones}
\usage{
my_theme_ggplot2(axis.size = NULL, axis.title.size = NULL)
}
\arguments{
\item{axis.size}{Tamaño de los ejes.}

\item{axis.title.size}{Tamaño de la leyenda.}
}
\value{
Un grafico de ggplot con el tema predefinido.
}
\description{
Usa parámetros de ggplot predefinidos para una publicación científica, se basa en mis criterios y gustos
personales. solo ejes y letras grandes.
}
\examples{
#Para usar el tema predefinido
#library(ggplot2)
#data(iris)
#ggplot(iris, aes(x=Petal.Length, y=Petal.Width)) +
#geom_point() +
#geom_smooth(method=lm , color="red", fill="#69b3a2", se=TRUE) +
#my_theme_ggplot2(axis.size= 10, axis.title.size= 12)

}
