% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA.cero.R
\encoding{UTF-8}
\name{NA.cero}
\alias{NA.cero}
\title{Sustituir todos los NA por ceros.}
\usage{
NA.cero(x)
}
\arguments{
\item{x}{Vector o columna}
}
\value{
Na por cualquier valor en un vector o columna.
}
\description{
Sustituir todos los NA por ceros.
}
\examples{
iris$Sepal.Length[c(1,3,5,7,9)] <- NA
iris.sinNA = data.frame(sapply(iris,NA.cero))
}
