% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA.map.R
\encoding{UTF-8}
\name{NA.map}
\alias{NA.map}
\title{Grafico que muestra los datos faltantes en un dataframe}
\usage{
NA.map(
  obj,
  legend = TRUE,
  col = c("indianred", "dodgerblue"),
  main,
  y.cex = 0.8,
  x.cex = 0.8,
  y.labels,
  y.at,
  csvar = NULL,
  tsvar = NULL,
  rank.order = TRUE,
  margins = c(5, 5),
  ...
)
}
\arguments{
\item{obj}{Dataframe}

\item{legend}{Titulo predefinido}

\item{col}{Colores usados predefinidos}

\item{main}{main predefinido}

\item{y.cex}{tamaño de leyendas de eje y predefinido}

\item{x.cex}{tamaño de leyendas de eje x predefinido}

\item{y.labels}{leyendas eje y predefinido}

\item{y.at}{y.at predefinido}

\item{csvar}{csvar predefinido}

\item{tsvar}{tsvar predefinido}

\item{rank.order}{rank order predefinido}

\item{margins}{margenes predefinido}

\item{...}{mas argumentos predefinidos}
}
\value{
Gráfico de distribución de datos faltantes.
}
\description{
Explorar visualmente donde están los datos faltantes en las diferentes columnas del dataframe. Esta función fue tomada de la paqueteria Amelia, https://github.com/IQSS/Amelia.
}
\examples{
data("iris")
iris$Sepal.Length[c(1,3,5,7,9)] <- NA
iris$Petal.Length[c(7,24,21,5)] <- NA
NA.map(iris)
}
