% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA.quitar.R
\encoding{UTF-8}
\name{NA.quitar}
\alias{NA.quitar}
\title{Eliminar los NA de una columna o varias columnas a la vez}
\usage{
NA.quitar(data, columnas)
}
\arguments{
\item{data}{Dataframe.}

\item{columnas}{Columnas de las cuales se van a remover los NA.}
}
\value{
Un vector sin NA.
}
\description{
Quitar NA de una sola columna o de varias columnas a la vez.
}
\examples{
data("iris")
#Introducimos NAs aleatorios en el dataframe
iris$Sepal.Width[c(1,3,5,7,9)] <- NA
iris$Petal.Width[c(2,5,6,7,9)] <- NA
head(iris)
#quitamos la columna Sepal.Width con NA
iris.sinNA2 <- NA.quitar(iris, "Sepal.Width")
head(iris.sinNA2)
#quitamos las columnas con NA
iris.sinNA3 <- NA.quitar(iris, c("Petal.Width","Sepal.Width"))
head(iris.sinNA3)
}
