% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA.remover.R
\encoding{UTF-8}
\name{NA.remover}
\alias{NA.remover}
\title{Eliminar los NA de una columna o varias columnas a la vez}
\usage{
NA.remover(data, columnas)
}
\arguments{
\item{data}{Dataframe.}

\item{columnas}{Columnas de las cuales se van a remover los NA.}
}
\value{
Un vector sin NA.
}
\description{
Quitar NA de una sola columna o de varias columnas a la vez.
}
\examples{
data("iris")
iris$Sepal.Width[c(1,3,5,7,9)] <- NA
iris$Petal.Width[c(1,3,5,7,9)] <- NA
iris2<- NA.remover(iris, c("Petal.Width","Sepal.Width"))
iris3<- NA.remover(iris, "Sepal.Width")
}
