% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA.removerSpecial.R
\encoding{UTF-8}
\name{NA.removerSpecial}
\alias{NA.removerSpecial}
\title{Reemplazar caracteres especiales por NA}
\usage{
NA.removerSpecial(df, vec)
}
\arguments{
\item{df}{dataframe}

\item{vec}{vector que contiene los caracteres especiales que desea sustituir por NA}
}
\value{
Vector sin caracteres especiales.
}
\description{
Reemplaza los caracteres especiales en el dataframe por NA.
}
\examples{
data(iris)
iris[2,2]<-"@"
iris[3,2]<-"???"
iris<-NA.removerSpecial(iris,c("@","???"))
head(iris)
}
