% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.lks.grupos.R
\encoding{UTF-8}
\name{norm.lks.grupos}
\alias{norm.lks.grupos}
\title{Prueba de normalidad de Lilliefors (Kolmogorov-Smirnov) por grupos.}
\usage{
norm.lks.grupos(formula, datos)
}
\arguments{
\item{formula}{Formula iniciando por la variable numerica de nuestro interes. Puede tomar
forma de y ~ x+z. Se pueden usar 2 o más grupos.}

\item{datos}{Base de datos o dataframe.}
}
\value{
Una prueba Lilliefors-Kolmogorov-Smirnov por grupos, un valor de p y estadistico
asociado a la prueba.
}
\description{
Realiza la prueba de Lilliefors (Kolmogorov-Smirnov) para probar la normalidad de una
variable por grupos.
}
\examples{
#data(iris)
#norm.lks.grupos(Petal.Length~ Species, iris)
#iris$site <- c(rep("A", 75), rep("B", 75))
#norm.lks.grupos(Petal.Length~ Species+ site, iris)
}
