% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.shapiro.grupos.R
\encoding{UTF-8}
\name{norm.shapiro.grupos}
\alias{norm.shapiro.grupos}
\title{Prueba de normalidad de Shapiro-Wilk por grupos.}
\usage{
norm.shapiro.grupos(formula, datos)
}
\arguments{
\item{formula}{Formula iniciando por la variable numerica de nuestro interes. Puede tomar forma de y ~ x+z. Se pueden usar 2 o más grupos.}

\item{datos}{Base de datos o dataframe.}
}
\value{
Una prueba Shapiro-Wilk por grupos
}
\description{
Realiza la prueba de normalidad de Shapiro-Wilk por grupos de un dataframe. Se pueden seleccionar una o más variables a la vez.
}
\examples{
#data(iris)
# prueba de normalidad Shapiro-Wilk con una variable de agrupamiento
#norm.shapiro.grupos(Petal.Length~ Species, iris)
# prueba de normalidad Shapiro-Wilk con dos o mas variables de agrupamiento
#iris$site <- c(rep("A", 75), rep("B", 75))
#norm.shapiro.grupos(Petal.Length~ Species+ site, iris)
}
