% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier.find.R
\encoding{UTF-8}
\name{outlier.find}
\alias{outlier.find}
\title{Encuentra los outliers de un dataframe}
\usage{
outlier.find(data, cutoff = 3)
}
\arguments{
\item{data}{dataframe}

\item{cutoff}{Parametro de tolerancia para oultiers.}
}
\value{
Una lista de los outliers.
}
\description{
Encuentra los outliers de un dataframe.
}
\examples{
outlier.find(iris)
outliers <- outlier.find(iris)
}
