% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.DHARMa.R
\encoding{UTF-8}
\name{outliers.DHARMa}
\alias{outliers.DHARMa}
\title{Encontrar outliers de un modelo con paqueteria DHARMa}
\usage{
outliers.DHARMa(Modelo)
}
\arguments{
\item{Modelo}{Modelo con lme4, glmer, glmmTMB, etc. GLMM y GLM.}
}
\value{
Un grafico de outliers y una lista con los outliers detectados.
}
\description{
Encontrar outliers con paqueteria DHARMa.
}
\examples{
#datos <- datasets::ChickWeight
#library(lme4)
#modelo <- glmer(weight ~ Diet +(1|Chick), family=gaussian("log"), data = datos)
#outliers.DHARMa(modelo)
}
