% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.glm.R
\encoding{UTF-8}
\name{outliers.glm}
\alias{outliers.glm}
\title{Encontrar outliers de un GLM}
\usage{
outliers.glm(model, dataframe)
}
\arguments{
\item{model}{Modelo con GLM y LM.}

\item{dataframe}{Dataframe con el cual se ajustó el modelo.}
}
\value{
Un grafico de outliers y una lista con los outliers detectados.
}
\description{
Encontrar outliers de un GLM. Usa las distancias de Cook e identifica los outliers. Ver la funcion cooks.distance, para mas detalles.
}
\examples{
data(iris)
modelo <- glm(Petal.Width ~ Petal.Length, family = gaussian("log"), data=iris)
outliers.glm(modelo, iris)
}
