% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.lme.R
\encoding{UTF-8}
\name{outliers.lme}
\alias{outliers.lme}
\title{Encontrar outliers de un modelo LMM}
\usage{
outliers.lme(modelo, random = NULL)
}
\arguments{
\item{modelo}{Modelo ajustado con la función lme de paqueteria nlme.}

\item{random}{Componente aleatorio del modelo a graficar.}
}
\value{
encontrar los outliers de un modelo lme.
}
\description{
Encontrar outliers de un LMM Linear mixed model de la paqueteria nlme, basandonos en las distancias de
Cook.
}
\examples{
#datos <- datasets::ChickWeight
#library(nlme)
#modelo <- lme(weight ~ Diet, random= ~1|Chick, data = datos)
#outliers.lme(modelo)
#outliers.lme(modelo, random = "Chick")
}
