% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.plot.R
\encoding{UTF-8}
\name{outliers.plot}
\alias{outliers.plot}
\title{Encontrar los outliers de un GLM, GLMM.}
\usage{
outliers.plot(model, outliers = 3)
}
\arguments{
\item{model}{Modelo GLM, GLMM, etc.}

\item{outliers}{Numero de outliers a mostrar.}
}
\value{
Encontrar los outliers de un modelo y graficarlos.
}
\description{
Encontrar los outliers de un GLM, GLMM O cualquier otro modelo. Además aplica una prueba para saber si
son o no datos influyentes.
}
\examples{
data(iris)
modelo <- glm(Petal.Width ~ Petal.Length, family = gaussian("log"), data=iris)
outliers.plot(modelo)
}
