% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.plot2.R
\encoding{UTF-8}
\name{outliers.plot2}
\alias{outliers.plot2}
\title{Encontrar los outliers de un GLM, LM}
\usage{
outliers.plot2(model, fill = "white", outline = "black", size = 30)
}
\arguments{
\item{model}{Modelo GLM, o LM.}

\item{fill}{color}

\item{outline}{outline}

\item{size}{tamaño}
}
\value{
Encontrar los outliers de un modelo y graficarlos.
}
\description{
Encontrar los outliers de un GLM.
}
\examples{
data(iris)
modelo <- glm(Petal.Width ~ Petal.Length, family = gaussian("log"), data=iris)
outliers.plot2(modelo)
}
