% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.plot3.R
\encoding{UTF-8}
\name{outliers.plot3}
\alias{outliers.plot3}
\title{Encontrar los outliers de un GLM, LM}
\usage{
outliers.plot3(model)
}
\arguments{
\item{model}{Modelo GLM, GLMM o LM.}
}
\value{
Encontrar los outliers de un modelo y graficarlos.
}
\description{
Encontrar los outliers de un GLM, GLMM, etc.
}
\examples{
data(iris)
modelo <- glm(Petal.Width ~ Petal.Length, family = gaussian("log"), data=iris)
outliers.plot3(modelo)
}
