% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe}
\usage{
lhs \%>\% rhs
}
\arguments{
\item{lhs}{Un valor o el marcador de posicion magrittr.}

\item{rhs}{Una llamada a una función que utiliza la semantica de magrittr.}
}
\value{
El resultado de llamar a \code{rhs(lhs)}.
}
\description{
Ver \code{magrittr::\link[magrittr:pipe]{\%>\%}} para mas detalles.
}
\keyword{internal}
