% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_effects.R
\encoding{UTF-8}
\name{plot_effects}
\alias{plot_effects}
\title{Graficar effectos de un LM, LMM, GLM, GLMM, GAM}
\usage{
plot_effects(modelo)
}
\arguments{
\item{modelo}{Modelo LM, GLM, GLMM, GAM, etc.}
}
\value{
Grafica todas las variables explicativas del modelo.
}
\description{
Graficar effectos de un LM, LMM, GLM, GLMM, GAM, etc. Permite hacer una observación rápida de las variables explicativas en el modelo y su influencia en la variable de respuesta. Gráficos generados con paqueteria "effects".
}
\examples{
data(iris)
modelo <- glm(Petal.Width ~ Petal.Length*Species+ Sepal.Length, family = gaussian("log"), data=iris)
plot_effects(modelo)
}
