% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_effects3.R
\encoding{UTF-8}
\name{plot_effects3}
\alias{plot_effects3}
\title{Graficar effectos de un LM, LMM, GLM, GLMM}
\usage{
plot_effects3(modelo, show.data = NULL)
}
\arguments{
\item{modelo}{Modelo LM, GLM, GLMM, etc.}

\item{show.data}{Argumento logico TRUE o FALSE, que indica si se agregan los puntos de las observaciones.}
}
\value{
Grafica todas las variables explicativas del modelo.
}
\description{
Graficar effectos de un LM, LMM, GLM, GLMM, etc. Permite hacer una observación rápida de las variables. Gráficos generados con paqueteria "Sjplot".
explicativas en el modelo y su influencia en la variable de respuesta.
}
\examples{
data(iris)
modelo <- glm(Petal.Width ~ Petal.Length*Species+ Sepal.Length, family = gaussian("log"), data=iris)
plot_effects3(modelo)
plot_effects3(modelo, show.data = TRUE)
plot_effects3(modelo, show.data = FALSE)
}
