% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid_DHARMa.R
\encoding{UTF-8}
\name{resid_DHARMa}
\alias{resid_DHARMa}
\title{Grafico de residuales de un modelo LM, LMM, GLM, GLMM, GAM, etc}
\usage{
resid_DHARMa(Modelo)
}
\arguments{
\item{Modelo}{Un modelo LM, GLM, GLMM, GAM,}
}
\value{
Grafica los residuales de un modelo.
}
\description{
Muestra el ajuste del modelo LM, GLM, GLMM, GAM, etc. Esta función crea residuos escalados
mediante la simulación del modelo ajustado. Los residuos pueden ser extraídos con residuals.DHARMa. Ver
testResiduals para una visión general de las pruebas de residuos, plot.DHARMa para una visión general de
los gráficos disponibles. Esta función fue creada para explorar el ajuste, normalidad y homogeneidad de
varianzas entre los grupos.
}
\examples{
#datos <- datasets::ChickWeight
#library(lme4)
#modelo <- glmer(weight ~ Diet +(1|Chick), family=gaussian("log"), data = datos)
#resid_DHARMa(modelo)
}
