% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid_glm.R
\encoding{UTF-8}
\name{resid_glm}
\alias{resid_glm}
\title{Grafico de residuales de un modelo GLM}
\usage{
resid_glm(Modelo)
}
\arguments{
\item{Modelo}{Un modelo LM, GLM.}
}
\value{
Esta función fue creada para explorar el ajuste, normalidad y homogeneidad de varianzas de un modelo GLM.
}
\description{
Muestra el ajuste del modelo LM, GLM. Esta función permite la exploración visual de los
residuales de un modelo.
}
\examples{
data(iris)
modelo <- glm(Petal.Width ~ Petal.Length, family = gaussian("log"), data=iris)
resid_glm(modelo)
}
