% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapiro_grupos.R
\encoding{UTF-8}
\name{shapiro_grupos}
\alias{shapiro_grupos}
\title{Prueba de normalidad de Shapiro-Wilk por grupos.}
\usage{
shapiro_grupos(formula, datos)
}
\arguments{
\item{formula}{Formula iniciando por la variable numerica de nuestro interes. Puede tomar forma de y ~ x+z. Se pueden usar 2 o más grupos.}

\item{datos}{Base de datos o dataframe.}
}
\value{
Una prueba Shapiro-Wilk por grupos
}
\description{
Realiza la prueba de normalidad de Shapiro-Wilk por grupos de un dataframe. Se pueden seleccionar una o más variables a la vez.
}
\examples{
#data(iris)
#shapiro_grupos(Petal.Length~ Species, iris)
#iris$site <- c(rep("A", 75), rep("B", 75))
#shapiro_grupos(Petal.Length~ Species+ site, iris)
}
