% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_col.R
\encoding{UTF-8}
\name{t_col}
\alias{t_col}
\title{Ajustar los colores a cierto porcentaje de transparencia}
\usage{
t_col(color, porcentaje = NULL, nombre = NULL)
}
\arguments{
\item{color}{Un color soportado por R. (p.ej. #66CD007F)}

\item{porcentaje}{Porcentaje de transparencia.}

\item{nombre}{Nombre opcional del color}
}
\value{
quitar ceros del vector.
}
\description{
Ajustar los colores a cierto porcentaje de transparencia.
}
\examples{
color_transparente <- t_col("#66CD007F", porcentaje = 50, nombre = "lt.color")
color_transparente
}
