% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_models.R
\encoding{UTF-8}
\name{table.models}
\alias{table.models}
\title{Exportar tabla de un modelo}
\usage{
table.models(...)
}
\arguments{
\item{...}{Uno o más modelos LM, GLM, GLMM, GAM, LMM, etc.}
}
\value{
Tabla del summary del modelo, con formato agradable lista para publicación.
}
\description{
Extraer el summary de un modelo LM, GLM, GLMM, GAM, etc. Generar una tabla en HTML copiable en Word.
}
\examples{
#table.models(model)
#table.models(m,m2,m3)
}
