% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_ANOVA2.R
\encoding{UTF-8}
\name{table_ANOVA2}
\alias{table_ANOVA2}
\title{Extraer tabla de ANOVA Tipo II}
\usage{
table_ANOVA2(modelo)
}
\arguments{
\item{modelo}{Un modelo GLM, GLMM, LMM, etc.}
}
\value{
Tabla ANOVA tipo II similar a la que se obtiene con la paqueteria car::Anova.
}
\description{
Extraer tabla de ANOVA Tipo II de un modelo. Extraer la tabla como HTML con formato publicable. Puede ser que algunos modelos no sean soportados y deberían llamarse directamente con la función Anova de la paquetería car.
}
\examples{
data(iris)
modelo <- glm(Petal.Width ~ Petal.Length, family = gaussian("log"), data=iris)
table_ANOVA2(modelo)
}
