% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_contrasts.R
\encoding{UTF-8}
\name{table_contrasts}
\alias{table_contrasts}
\title{Exportar tabla de contrastes post hoc}
\usage{
table_contrasts(cont.emmeans, title = NULL, sort = FALSE)
}
\arguments{
\item{cont.emmeans}{Contrastes hechos con emmeans de un modelo LM,GLM, GLMM, GAM, LMM, etc.}

\item{title}{Titulo opcional}

\item{sort}{Argumento TRUE o FALSE, permite decidir si se ordenan los valores de P.}
}
\value{
Tabla de contrastes.
}
\description{
Extraer los contrastes de un modelo, para generar una tabla en HTML copiable en Word. Para usar
esta función es necesario crear un objeto con los contrastes previamente hechos con la paqueteria
emmeans.
}
\examples{
data(iris)
modelo <- glm(Petal.Width ~ Petal.Length+Species, family = gaussian("log"), data=iris)
library(emmeans)
cont <- emmeans(modelo,pairwise ~ Species,adjust="tukey",type="response")$contrasts
table_contrasts(cont)
#PARA ORDENAR DE MENOR A MAYOR P-VALUE:
table_contrasts(cont, sort = TRUE)
}
