% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temas_ggplot.R
\encoding{UTF-8}
\name{tema_articulo}
\alias{tema_articulo}
\title{#' Tema 1 de ggplot predefinidos por mi para publicaciones}
\usage{
tema_articulo(axis.size = NULL, axis.title.size = NULL)
}
\arguments{
\item{axis.size}{Tamaño de los ejes.}

\item{axis.title.size}{Tamaño de la leyenda del eje.}
}
\value{
Un grafico de ggplot con el tema predefinido.
}
\description{
Usa parámetros de ggplot predefinidos para una publicación científica, se basa en mis criterios y gustos
personales. Se puede especificar tamaño de los ejes X y Y, negritas y sin fondo.
}
\examples{
#Para usar el tema predefinido
#require(ggplot2)
#data(iris)
#ggplot(iris, aes(x=Petal.Length, y=Petal.Width)) +
#geom_point() +
#geom_smooth(method=lm , color="red", fill="#69b3a2", se=TRUE) +
#tema_articulo(axis.size= 10, axis.title.size= 12)
}
