% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.arcsine.R
\encoding{UTF-8}
\name{trans.arcsine}
\alias{trans.arcsine}
\title{Transformar variable con arcsine, proporciones}
\usage{
trans.arcsine(x)
}
\arguments{
\item{x}{vector que contiene las proporciones en escalas de porcentaje de 0-100.}
}
\value{
vector transformado.
}
\description{
Permite la transformación de Arcsine aplicada a proporciones. Aplicar la transformación arcsine se ha
sugerido por mucho tiempo como un método estandar al tratar con proporciones/porcentajes (Sokahl & Rohlf, #' 1995). Ver: Sokahl, R., & Rohlf, F. (1995). Biometry (3rd ed.). New York: W. H. Freeman.
}
\examples{
#variable <- seq(1,99, 3)
#trans.arcsine(variable)
}
