% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.binomial.R
\encoding{UTF-8}
\name{trans.binomial}
\alias{trans.binomial}
\title{Convertir en 0 y 1 un vector numerico}
\usage{
trans.binomial(dataframe, variable)
}
\arguments{
\item{dataframe}{dataframe}

\item{variable}{variable numerica a transformar en binomial.}
}
\value{
vector transformado en 0 y 1.
}
\description{
Permite convertir cualquier variable numerica en variable binomial de 0 y 1, presencia-ausencia.
}
\examples{
#df <- trans.binomial(dataframe = data, variable = concentration)
}
