% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.logit.R
\encoding{UTF-8}
\name{trans.logit}
\alias{trans.logit}
\title{Transformar variable con logit}
\usage{
trans.logit(p)
}
\arguments{
\item{p}{vector que contiene las proporciones en escalas de porcentaje de 0-100.}
}
\value{
vector transformado.
}
\description{
Permite la transformación de logit aplicada a proporciones. Aplicar la transformación logit seguida de una regresión por mínimos cuadrados no lineales sobre  las variables transformadas (Warton & Hui, 2011). Este metodo se ha sugerido por mucho tiempo como un  método para tratar con proporciones/porcentajes. Ver: Warton, D. I., & Hui, F. K. C. (2011). The arcsine is asinine: The analysis of proportions in ecology. Ecology, 92(1), 3–10. https://doi.org/10.1890/10‐0340.1
}
\examples{
variable <- seq(1,99, 3)
trans.logit(variable)
}
