% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unscale.visreg.xaxis.R
\encoding{UTF-8}
\name{unscale.visreg.xaxis}
\alias{unscale.visreg.xaxis}
\title{Desescalar coeficientes previamente escalados y centrados, para graficar con visreg}
\usage{
unscale.visreg.xaxis(x)
}
\arguments{
\item{x}{var}
}
\value{
coeficientes o medias reconvertidas en su escala original
}
\description{
Permite dibujar las regresiones parciales de un modelo usando la paqueteria visreg. Permite desescalar y descentrar los coeficientes de un modelo ajustado con las variables escaladas. Para convertir la variable del eje x a la escala original. Solo hay que sustituir la "variable"  por la variable en escala original. Luego incluir en la funcion de visreg xtrans = unscale.visreg.xaxis. Se usa para graficar en visreg.
}
\examples{
#Esta función se usa cuando se quiere graficar los efectos de una variable escalada.
#visreg(m2, xtrans = unscale.visreg.xaxis)
}
