% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vacio.as.NA.R
\encoding{UTF-8}
\name{vacio.as.NA}
\alias{vacio.as.NA}
\title{Convierte celdas vacías en NA}
\usage{
vacio.as.NA(x)
}
\arguments{
\item{x}{Dataframe con los datos faltantes no codificados como NA.}
}
\value{
Un vector con NA.
}
\description{
Datos faltantes en blanco por NA.
}
\examples{
library(dplyr)
A  <- data.frame(x=c(1,2,3,4),y=c(1,2,3,4), z=c(1,"N/A",3,4))
A \%>\% mutate_each(funs(vacio.as.NA))
}
