% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.distr.R
\encoding{UTF-8}
\name{var.distr}
\alias{var.distr}
\title{Graficar la distribucion de una variable numerica.}
\usage{
var.distr(variable)
}
\arguments{
\item{variable}{Variable de interes.}
}
\value{
gráfico con la distribución más probable de la variable.
}
\description{
Permite conocer la distribución de una variable numerica. Usando la paqueteria fitdistrplus, ajusta las distribuciones más probables de un vector. Para tener en cuenta la incertidumbre de los valores estimados de la curtosis y la asimetría a partir de los datos, el conjunto de datos puede someterse a un proceso de bootstrap fijando el argumento boot en un número entero superior a 10. A continuación, se calculan los valores de asimetría y curtosis correspondientes a las muestras de bootstrap y se presentan en color azul en el gráfico de asimetría y curtosis.
}
\examples{
#x<- rnorm(1000,313, 5.7)
#var.distr(x)
#var.distr(datos$variable)
}
