% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.fitdistr.R
\encoding{UTF-8}
\name{var.fitdistr}
\alias{var.fitdistr}
\title{Ajustar una variable a cierta distribución de nuestro interes.}
\usage{
var.fitdistr(variable, distr)
}
\arguments{
\item{variable}{Variable de interes.}

\item{distr}{Distribución a probar: "norm", "lnorm", "pois", "exp", "gamma", "nbinom", "geom", "beta",
"unif" and "logis"}
}
\value{
coeficientes o medias reconvertidas en su escala original
}
\description{
Permite conocer ajustar una variable a cierta distribución de nuestro interes.
}
\examples{
#x<- rnorm(1000,313, 5.7)
#var.fitdistr(x, distr = "norm")
}
