% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VIF.R
\encoding{UTF-8}
\name{VIF}
\alias{VIF}
\title{Obtener el VIF de un dataframe solo tomando en cuenta variables numéricas}
\usage{
VIF(df)
}
\arguments{
\item{df}{Un dataframe.}
}
\value{
Lista de las variables y su correspondiente VIF.
}
\description{
Calcule Factores de inflación de la variación (VIF) para estimar la multi-colinearidad
de las variables de un dataframe. Cuando al ajustar un modelo, arroja un error: Error in solve.default(as.matrix(fit$hessian)) : system is computationally singular: reciprocal condition number = 4.31555e-18 stepAIC). Se recomienda quitar las variables si el VIF es mayor a 10, o siendo más estrictos mayor a 3.
}
\examples{
VIF(iris)
VIF(mtcars)
}
