% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VIF_model.R
\name{VIF_model}
\alias{VIF_model}
\alias{multicollinearidad}
\title{Revisar la multicolinealidad de las variables del modelo}
\usage{
VIF_model(x, ...)

multicollinearidad(x, ...)
}
\arguments{
\item{x}{Un modelo. Actualmente acepta lm, glm, lmm y glmmm (hasta donde yo sé).}

\item{...}{Actualmente no se utiliza.}
}
\value{
Un dataframe con información sobre el nombre del término del modelo, el
VIF y los intervalos de confianza asociados, el factor
por el que se incrementa el error estándar debido a la posible correlación
con otros términos, y los valores de tolerancia (incluyendo los intervalos de confianza),
donde \code{tolerance = 1/vif}.
}
\description{
\code{VIF_model()} comprueba la multicolinealidad en los modelos de regresión
intervals for VIF and tolerance are based on Marcoulides et al.
(2019, Appendix B).
\code{check_concurvity()} es una envoltura alrededor de \code{mgcv::concurvity()}, y puede ser
considerado como una comprobación de colinealidad para los términos 'suavizados' en los
GAM. Confidence intervals for VIF and tolerance are based on Marcoulides et al.
(2019, Appendix B).
}
\note{
El código para calcular los intervalos de confianza para el VIF y la tolerancia
fue adaptado del Apéndice B del artículo de Marcoulides et al.
Por lo tanto, los créditos van a estos autores el algoritmo original. También
a \href{https://easystats.github.io/see/articles/performance.html}{\code{plot()}-method}
implementado en el \href{https://easystats.github.io/see/}{\pkg{see}-paquete}.
}
\examples{
m <- lm(mpg ~ wt + cyl + gear + disp, data = mtcars)
VIF_model(m)

# plot results
if (require("see")) {
  x <- VIF_model(m)
  plot(x)
}
}
\references{
\itemize{
\item Marcoulides, K. M., and Raykov, T. (2019). Evaluation of Variance
Inflation Factors in Regression Models Using Latent Variable Modeling
Methods. Educational and Psychological Measurement, 79(5), 874–882.
}
}
