% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq.post.hoc.R
\encoding{UTF-8}
\name{chisq.post.hoc}
\alias{chisq.post.hoc}
\title{Contrastes post hoc de una matriz de datos con chi-cuadrada}
\usage{
chisq.post.hoc(
  tbl,
  test = c("fisher.test"),
  popsInRows = TRUE,
  control = c("fdr", "BH", "BY", "bonferroni", "holm", "hochberg", "hommel"),
  digits = 4,
  simulate.p.value = NULL
)
}
\arguments{
\item{tbl}{Una tabla o matriz usada para la función chisq.test}

\item{test}{Método usado para ajustar el valor de P.}

\item{popsInRows}{Un argumento lógico que indica si las poblaciones forman las filas (por defecto; =TRUE
) de la tabla o no (=FALSE).}

\item{control}{Indica el método de control a utilizar.}

\item{digits}{Número de digitos a mostrar.}

\item{simulate.p.value}{Si argumento es TRUE, indica que se calculan los valores p mediante la simulación de Monte Carlo y 5000 replicas. Para más información consulta la funcion 'chisq.test'.}
}
\value{
Contrastes 'post-hoc' de chi cuadrada para dos niveles usando Fisher.test.
}
\description{
Muestra las comparaciones post-hoc por pares de una prueba Chi cuadrada. Calcula los contrastes si hay
más de dos grupos en una chisq. Las pruebas post-hoc para saber que pares de poblaciones difieren tras
una prueba de chi-cuadrado significativa pueden construirse realizando todas las pruebas de chi-cuadrado
para todos los pares de poblaciones y ajustando después los valores p resultantes. Los valores p
ajustados pueden calcularse con una amplia variedad de métodos: fdr, BH, BY, bonferroni, holm, hochberg y hommel. Funciona básicamente como una función envolvente que envía los valores p "brutos" no ajustados de cada prueba de chi-cuadrado por pares a la función p.adjust en el programa R base. La función p.adjust debe ser consultada para una mayor descripción de los métodos utilizados. Función tomada de la paqueteria fifer. https://www.rdocumentation.org/packages/fifer/versions/1.1
}
\examples{
table <- as.table(rbind(c(14, 43), c(23, 47)))
rownames(table) <- c("females", "males")
chisq.post.hoc(table,test = c("fisher.test"))
# con más filas:
M <- as.table(rbind(c(762, 327, 468), c(484, 239, 477)))
dimnames(M) <- list(gender = c("F", "M"), party = c("Especie1","Especie2", "Especie3"))
chisq.post.hoc(M,test = c("fisher.test"))
}
