% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispfun2.R
\encoding{UTF-8}
\name{dispfun2}
\alias{dispfun2}
\title{Parametro de dispersion de un modelo poisson, o binomial negativo}
\usage{
dispfun2(modelo)
}
\arguments{
\item{modelo}{Modelo con distribuci칩n Poisson, Quasipoisson, binomial negativo.}
}
\value{
Un valor de dispersi칩n.
}
\description{
Calcular parametro de dispersi칩n para poisson, binomial negativo y Quasipoison de la paqueteria lme4,
glmer, glmmTMB, etc. GLMM y GLM. M치s detalles en:  https://github.com/glmmTMB/glmmTMB/issues/224
}
\examples{
data<- warpbreaks
modelo <- glm(breaks ~ wool+tension, poisson, data= data)
dispfun2(modelo)
}
