% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpairs_dfnum.R
\encoding{UTF-8}
\name{ggpairs_dfnum}
\alias{ggpairs_dfnum}
\title{ggpairs_dfnum}
\usage{
ggpairs_dfnum(dataframe)
}
\arguments{
\item{dataframe}{Dataframe}
}
\value{
Un gráfico de correlación con valores de correlación, valores de p y lineas de
regresión lineal y no lineal de cada relación.
}
\description{
Matriz de correlación basada en ggpairs con lineas de regresion lineal y no lineal. Matriz de correlación de variables numéricas. Esta función selecciona solo las variables numéricas y aplica la función ggpairs
de la paqueteria GGally.
}
\examples{
ggpairs_dfnum(iris)
}
