% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_ggplot_themes2.R
\encoding{UTF-8}
\name{my_theme_ggplot2}
\alias{my_theme_ggplot2}
\title{Tema2  de ggplot predefinidos por mi para publicaciones}
\usage{
my_theme_ggplot2(axis.size = NULL, legend.size = NULL)
}
\arguments{
\item{axis.size}{Tamaño de los ejes}

\item{legend.size}{Tamaño de la leyenda}

\item{title.plot}{Titulo del gráfico}
}
\value{
Un grafico de ggplot con el tema predefinido.
}
\description{
Usa parámetros de ggplot predefinidos para una publicación científica, se basa en mis criterios y gustos
personales. solo ejes y letras grandes.
}
