% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.GLMM.R
\encoding{UTF-8}
\name{optimizeGLMM}
\alias{optimizeGLMM}
\title{Probar optimizadores en modelo glmmTMB}
\usage{
optimizeGLMM(model, optimizer = c("optim", "nlminb"))
}
\arguments{
\item{model}{Modelo GLMM ajustado con la paqueteria glmmTMB.}

\item{optimizer}{Optimizador que se usa en el modelo, puede ser "optim" o "nlminb".}
}
\value{
Lista con los diferentes optimizadores y su ajuste. Si es FALSE, el modelo no converge.
}
\description{
Esta es una función que permite ajustar GLMM con diferentes optimizadores cuando existen problemas
de convergencia.
}
\examples{
# do not run
#datos <- datasets::ChickWeight
#library(glmmTMB)
#modelo <- glmmTMB(weight ~ Diet +(1|Chick), family=gaussian("log"), data = datos)
# optimizeGLMM(modelo, optimizer = "optim")
# optimizeGLMM(modelo, optimizer = "nlminb")
}
