% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.GLMM.R
\encoding{UTF-8}
\name{optimizeGLMM}
\alias{optimizeGLMM}
\title{Probar optimizadores en modelo glmmTMB}
\usage{
optimizeGLMM(modelo)
}
\arguments{
\item{modelo}{Modelo GLMM ajustado con la paqueteria glmmTMB.}
}
\value{
Lista con los diferentes optimizadores y su ajuste. Si es FALSE, el modelo no converge.
}
\description{
Esta es una función que permite ajustar GLMM con diferentes optimizadores cuando existen problemas de convergencia. A veces al ajustar modelos pueden arrojar errores. Por ejemplo, una advertencia ‘iteration limit reached without convergence’, se soluciona aumentando el número de iteraciones: 'glmmTMBControl(optCtrl=list(iter.max=1e3,eval.max=1e3))'. Vea la ayuda de la función glmmTMBControl para mas detalles.
}
\examples{
# do not run
#datos <- ChickWeight
#library(glmmTMB)
#modelo <- glmmTMB(weight ~ Diet +(1|Chick), family=gaussian("log"), data = datos)
# optimizeGLMM(modelo)
}
