% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.DHARMa.R
\encoding{UTF-8}
\name{outliers.DHARMa}
\alias{outliers.DHARMa}
\title{Encontrar outliers de un modelo con paqueteria DHARMa}
\usage{
outliers.DHARMa(Modelo, nsim = NULL)
}
\arguments{
\item{Modelo}{Modelo con lme4, glmer, glmmTMB, etc. GLMM, GLM y GAM.}

\item{nsim}{Número de simulaciones. Cuanto menor sea el número, mayor será el error estocástico en los residuales.
Además, para n muy pequeño, los artefactos de discretización pueden influir en las pruebas. Por defecto es 250, que
es un valor relativamente seguro. Puede considerar aumentar a 1000 para estabilizar los valores simulados.}
}
\value{
Un grafico de outliers y una lista con los outliers detectados.
}
\description{
Encontrar outliers con paqueteria DHARMa. Esta función comprueba si el número de observaciones fuera de la simulación es mayor o menor de lo esperado. Ver función "testOutliers" de la paquetería DHARMa para más detalles.
}
\examples{
#datos <- datasets::ChickWeight
#library(lme4)
#modelo <- glmer(weight ~ Diet +(1|Chick), family=gaussian("log"), data = datos)
#outliers.DHARMa(modelo)
}
