% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.data.R
\encoding{UTF-8}
\name{scale_data}
\alias{scale_data}
\title{Estandarizar las variables de un dataframe, excluyendo algunas}
\usage{
scale_data(df, ex_variables, center = TRUE, scale = TRUE)
}
\arguments{
\item{df}{Un dataframe}

\item{ex_variables}{Variables a excluir del dataframe. Esto si no queremos aplicar la escalada a dichas variables.}

\item{center}{centrado}

\item{scale}{escalado}
}
\value{
vector escalado
}
\description{
Estandariza las todas las variables de un dataframe mediante el centrado y el escalado. Ver stdize en MuMIn package para detalles.
}
\examples{
data(iris)
df.sc <- scale_data(iris, ex_variables = c("Petal.Width","Petal.Length"))
}
