% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_models.R
\encoding{UTF-8}
\name{table.models}
\alias{table.models}
\title{Exportar tabla de un modelo}
\usage{
table.models(..., transformar = NULL)
}
\arguments{
\item{...}{Uno o más modelos LM, GLM, GLMM, GAM, LMM, etc.}

\item{transformar}{Si es TRUE, se aplicara la transformacion inversa de la funcion de liga del modelo. Si no se proporciona un argumento, se mostraran los estimados y el std.error en la escala de la función de liga. La transformación solo se aplica al primer modelo, no se puede aplicar a multiples modelos a la vez.}
}
\value{
Tabla del summary del modelo, con formato agradable lista para publicacion.
}
\description{
Extraer el summary de un modelo LM, GLM, GLMM, GAM, etc. Generar una tabla en HTML copiable en Word.
}
\examples{
# muestra valores sin transformar en la escala de la funcion de liga
#table.models(model)
#muestra valores sin transformar de todos los modelos
#table.models(m,m2,m3)
# muestra valores re-transformados a la escala original
#table.models(model, transformar=TRUE)
# muestra valores sin transformar
#table.models(model, transformar=FALSE)
}
