% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.01.beta.R
\encoding{UTF-8}
\name{trans.01.beta}
\alias{trans.01.beta}
\title{Transformar variable para ajustar regresión beta, util para proporciones}
\usage{
trans.01.beta(x)
}
\arguments{
\item{x}{vector que contiene las proporciones en escalas de porcentaje entre 0 y 1.}
}
\value{
vector transformado.
}
\description{
Cuando se tienen 0 y 1 absolutos en las proporciones, no se puede ajustar un modelo de regresion Beta. Arroja una advertencia: invalid dependent variable, all observations must be in (0, 1). Se recomienda aplicar la transformacion de las proporciones para ajustar modelos con familia beta, propia para proporciones entre 0 y 1. Esta transformacion permite lidiar con los 0 y 1 absolutos (Douma, 2019).
}
\examples{
x<- seq(0.1, 0.9, 0.01)
trans.01.beta(x)
#trans.01.beta(df$proporciones)
}
\references{
\itemize{
\item Douma, JC, Weedon, JT. Analysing continuous proportions in ecology and evolution:
A practical introduction to beta and Dirichlet regression. Methods Ecol Evol. 2019; 10
:1412– 1430. https://doi.org/10.1111/2041-210X.13234
}
}
