% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unscale.coef.R
\encoding{UTF-8}
\name{unscale.coef}
\alias{unscale.coef}
\title{Desescalar coeficientes previamente escalados y centrados.}
\usage{
unscale.coef(coef, scaled_covariate)
}
\arguments{
\item{coef}{Coeficientes del modelo.}

\item{scaled_covariate}{variable escalada que contiene la media y la desviación estandar como atributos.}
}
\value{
coeficientes o medias reconvertidas en su escala original
}
\description{
Permite desescalar y descentrar los coeficientes de un modelo ajustado con las variables
escaladas.
}
\examples{
#unscale.coef(-0.3440, using_scale)
#unscale.coef(coefs, df_scaled)
}
